Changed data type for the water usage amounts
This change is used to accommodate floating point values which may appear in the future

+++ data_models/forecasts/__init__.py
@@ -18,7 +18,7 @@ class RealData(BaseModel):
         default=...,
         alias='timePeriodEnd'
     )
-    water_usage_amounts: List[int] = Field(
+    water_usage_amounts: List[float] = Field(
         default=...,
         alias='waterUsageAmounts'
     )
